Make RECT use int32_t instead of long

+++ src/platform/PlatformAbstraction.h
@@ -184,7 +184,7 @@ typedef uint16_t       WORD;
 
 typedef struct
 {
-	long left; long top; long right; long bottom;
+	int32_t left; int32_t top; int32_t right; int32_t bottom;
 } RECT;
 
 unsigned long timeGetTime(void);
